requirements.txt
SimpleITK==2.0.2
torch==1.10.2+cu102
torchaudio==0.10.2+cu102
torchsummary==1.5.1
torchvision==0.11.3+cu102
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
grpcio-tools==1.44.0
numpy==1.20.3
protobuf==3.19.1
pydicom==2.2.2
requests==2.26.0
requests-oauthlib==1.3.0
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX @ file:///tmp/build/80754af9/tensorboardx_1621440489103/work
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-io-gcs-filesystem==0.22.0